# specs/003-catalog-api/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Clinic Catalog API
  version: 1.2.0
  description: >
    Public, read-only catalog API (Feature 003). Single-tenant: every clinic value comes from the
    database. Field names and types match frontend/src/types/content.ts; fields marked
    "additive" are new and optional for existing clients. All prices are whole PKR. Schedule
    times are wall-clock "HH:MM" in the clinic's timeZone.

    Every response carries X-Request-ID and security headers. Catalog 200 responses carry
    ETag and Cache-Control: public, max-age=300; a matching If-None-Match returns 304.
    Every /api/v1 endpoint is rate limited per IP (default 60/min) and may return 429.

    Availability and booking (Feature 005). Instants are UTC ISO 8601 ("Z"); every instant is
    accompanied by localDate / localTime in the clinic timeZone. Slot and booking responses are
    never cached (Cache-Control: no-store). POST /appointments is a server-to-server call: it
    requires X-Proxy-Secret, and X-Client-IP is honoured only with a valid secret.

    Command Centre (Feature 006). Every /api/v1/admin/* operation (tag command-centre) is called only by the
    website server: it requires X-Proxy-Secret, all but sign-in and demo start also require X-Session-Token,
    and state-changing operations (and POST /admin/bookings/search) require X-CSRF-Token. Responses are
    Cache-Control: no-store and X-Robots-Tag: noindex. The access policy of each operation is `x-policy`
...
servers:
...
paths:
...
  /api/v1/admin/staff:
    get:
      tags: [command-centre]
      operationId: adminStaffList
      x-policy: READ_ADMIN
      security: [ { proxySecret: [], sessionToken: [] } ]
      responses:
        "200": { description: staff accounts (demo → synthetic), content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Staff" } } } } }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [command-centre]
      operationId: adminStaffCreate
      x-policy: WRITE_ADMIN
      security: [ { proxySecret: [], sessionToken: [], csrfToken: [] } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/StaffCreate" } } } }
      responses:
        "201": { description: created (must change password at first sign-in), content: { application/json: { schema: { $ref: "#/components/schemas/Staff" } } } }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { description: email_taken }
        "422": { $ref: "#/components/responses/Validation" }
  /api/v1/admin/staff/{staffId}/reset-password:
    post:
      tags: [command-centre]
      operationId: adminStaffResetPassword
      x-policy: WRITE_ADMIN
      security: [ { proxySecret: [], sessionToken: [], csrfToken: [] } ]
      parameters: [ { name: staffId, in: path, required: true, schema: { type: string, format: uuid } } ]
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [temporaryPassword], properties: { temporaryPassword: { type: string, minLength: 12, maxLength: 128 } } } } } }
      responses:
        "204": { description: reset; all of that person's sessions ended }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { description: not_found }
        "422": { $ref: "#/components/responses/Validation" }
  /api/v1/admin/staff/{staffId}:
    patch:
      tags: [command-centre]
      operationId: adminStaffUpdate
      x-policy: WRITE_ADMIN
      security: [ { proxySecret: [], sessionToken: [], csrfToken: [] } ]
      description: Change role and/or active flag. Deactivation ends all sessions. Refuses to leave zero active admins (409 last_admin).
      parameters: [ { name: staffId, in: path, required: true, schema: { type: string, format: uuid } } ]
      requestBody: { required: true, content: { application/json: { schema: { type: object, minProperties: 1, properties: { role: { $ref: "#/components/schemas/Role" }, isActive: { type: boolean } } } } } }
      responses:
        "200": { description: updated, content: { application/json: { schema: { $ref: "#/components/schemas/Staff" } } } }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { description: not_found }
        "409": { description: last_admin }
components:
  securitySchemes:
    proxySecret: { type: apiKey, in: header, name: X-Proxy-Secret }
    sessionToken: { type: apiKey, in: header, name: X-Session-Token }
    csrfToken: { type: apiKey, in: header, name: X-CSRF-Token }
...
  responses:
...
    Unauthenticated: { description: "401 not_signed_in / session_expired", content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
    Forbidden: { description: "403 forbidden / demo_read_only / password_change_required / csrf_failed", content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
...
    Validation: { description: "422 validation_error / weak_password", content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
...
  schemas:
...
    ErrorInfo:
      type: object
      required: [code, message, requestId]
      properties:
        code: { type: string, enum: [not_found, method_not_allowed, validation_error, rate_limited, internal_error, service_unavailable, not_configured, forbidden, request_rejected, slot_taken, slot_unavailable, booking_limit_reached, idempotency_key_reused, not_signed_in, session_expired, sign_in_failed, account_locked, csrf_failed, demo_read_only, password_change_required, weak_password, email_taken, last_admin, transition_not_allowed, booking_changed, undo_unavailable] }
        message: { type: string }
        requestId: { type: string }
        retryAfterSeconds: { type: integer }
        details:
          type: array
          items: { type: object, required: [field, issue], properties: { field: { type: string }, issue: { type: string } } }
...
    ErrorResponse:
      type: object
      required: [error]
      properties: { error: { $ref: "#/components/schemas/ErrorInfo" } }
    Role: { type: string, enum: [admin, receptionist] }
...
    Staff:
      type: object
      required: [id, email, displayName, role, isActive]
      properties:
        id: { type: string, format: uuid }
        email: { type: string }
        displayName: { type: string }
        jobTitle: { type: string, description: "Sample staff only: the role title shown on the demo staff list." }
        role: { $ref: "#/components/schemas/Role" }
        isActive: { type: boolean }
        mustChangePassword: { type: boolean }
        lastSignInAt: { type: [string, "null"], format: date-time }
        isSample: { type: boolean }
    StaffCreate:
      type: object
      required: [email, displayName, role, temporaryPassword]
      properties: { email: { type: string }, displayName: { type: string }, role: { $ref: "#/components/schemas/Role" }, temporaryPassword: { type: string, minLength: 12, maxLength: 128 } }
...